Make rel helper functions more abstract
This modifies append_desc_part() into append_exps_except(), which takes
a relation and a list of exps, and selects all exps of the relation that
are not present in the list of exps. And append_appl_part() becomes
append_exps(), which simple selects all exps. All selected exps are
appended to the input list.